stock_volatility_prediction/requirements.txt
# Core dependencies first
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
python-dotenv>=1.0.0
pyyaml>=6.0

# ML/Stats packages
scikit-learn>=1.3.0
scipy>=1.10.0
arch>=6.0.0

# Deep Learning (optional - can be removed if not needed)
tensorflow>=2.13.0,<2.20.0
keras>=3.0.0

# Data fetching
yfinance>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web framework
streamlit>=1.28.0

# MLflow and tracking
mlflow>=2.8.0
dagshub>=0.3.0
